chore(StateManager): replace `makeStateManager` by `StateManagerFactory` for consistency
