+++ packages/devextreme/js/__internal/core/state_manager/state_manager.ts
@@ -1,7 +1,11 @@
+import { Logger } from './logger';
+import { PreactSignalValueContainerManagerFactory } from './preact_signal_value_container_manager';
+// eslint-disable-next-line spellcheck/spell-checker
+import { ReduxDevToolsConnector } from './redux_dev_tools_connector';
 import type * as StateManagementTypes from './types';
 import { deepCopy, joinStatePath } from './utils';
 
-export class StateManager implements StateManagementTypes.StateManager {
+class StateManager implements StateManagementTypes.StateManager {
   private readonly devToolsConnector: StateManagementTypes.DevToolsConnector;
 
   private readonly logger: StateManagementTypes.Logger;
@@ -173,3 +177,23 @@ export class StateManager implements StateManagementTypes.StateManager {
     return result;
   }
 }
+
+export const StateManagerFactory = {
+  create: (options: StateManagementTypes.StateManagerFactoryOptions): StateManager => {
+    const logger = options.logger ?? new Logger({ logLevel: options.logLevel, prefix: '[StateManager]' });
+
+    const stateContainerManagers: StateManagementTypes.StateManagerConfig['valueContainerManagers'] = options.valueContainerManagers ?? [PreactSignalValueContainerManagerFactory];
+
+    const preparedConfig: StateManagementTypes.StateManagerConfig = {
+      valueContainerManagers: stateContainerManagers,
+      devToolsConnector: options.devToolsConnector
+          // eslint-disable-next-line spellcheck/spell-checker
+          ?? new ReduxDevToolsConnector(options.componentName, logger),
+      logger,
+      stateSourceSign: options.stateSourceSign,
+    };
+
+    return new StateManager(preparedConfig);
+  },
+
+};
